apps/design/admin.py
from django import contrib
from django.shortcuts import get_object_or_404
from .models import *
from portfolio import models
from sorl.thumbnail.admin import AdminImageMixin

from multiupload.admin import MultiUploadAdmin