requirements.txt
django-admin-multiupload==1.10 #pip install git+git://github.com/gkuhn1/django-admin-multiupload.git
asgiref==3.2.10
beautifulsoup4==4.9.1
Django==3.1
django-admin-tools==0.9.0
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-uuslug==1.2.0
Pillow==7.2.0
python-memcached==1.59
python-slugify==4.0.1
pytz==2020.1
six==1.15.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
Wand==0.6.2
